LPS-60842 Move CascadeFileEntryTypesCallable to a public class to avoid duplicate code
